Fixed the overflow bug
Fixes chennaipy/website#39

Avoid using `style="float: right"` inline CSS as it causes image overflow in the meeting minutes page. Also avoid using long strings in the alt attribute of the img tag as some part of the string is skipped causing the end quote to be skipped as well. This causes the the meta information(like tag, date) to be displayed on the same line as that of the image(see November meetup minutes in the website).